Adding auto updating to the application
- Added the auto_update function in mbaacc.py, which runs on init

auto_update first compares the current config.CURRENT_VERSION with the latest github release tag to see if it's an older version.

If it is an older version, it will download the latest concerto.exe from the github releases, close concerto.exe, replace it with the new one, and start concerto.exe.

Current issues still present: Windows only, no error handling in the batch file, no asking the user to update concerto just an instant update.